requirements.txt
fastapi==0.115.14
pydantic-settings==2.10.1
uvicorn==0.35.0
sqladmin==0.21.0
itsdangerous==2.2.0
PyYAML==6.0.2
dependency_injector==4.48.1
sqlalchemy[asyncio]==2.0.41
faststream[rabbit]==0.5.44
faststream[redis]==0.5.44
asyncpg==0.30.0
sentry-sdk==2.34.1
aiogram==3.22.0
prometheus-client==0.22.1
prometheus_fastapi_instrumentator==7.1.0
sqladmin==0.21.0
itsdangerous==2.2.0
